move MailBuilder out of Helper namespace
